Remove missing penalty for bannedLicense check
This check depends on the presence of a cycloneDX-format BOM file, as
does the depScanCheck. When the underlying tooling that generates this
BOM file errors-out and produces no BOM file, no depscan or license
checking can occur. This change makes it so that at least users are not
being penalized twice for the same missing file.